requirements.txt
fastapi
uvicorn
websockets
python-multipart
google-generativeai
Pillow
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv